TP_240314_1
Solucionado lo de las fechas falta convertirlas a minutos para hacer
los cálculos

+++ Parking/src/parking/Caja.java
@@ -17,7 +17,7 @@ public class Caja {
         return precio;
     }
     public Ticket crearTicket(int numeroTicket, Vehiculo v){
-       float tiempo;
+       double tiempo;
        
        
        tiempo = v.obtenerTiempo();
